add structs for different modes
