+++ packages/daylight/modes/vde.go
@@ -1,0 +1,15 @@
+package modes
+
+import "github.com/GenesisKernel/go-genesis/packages/conf"
+
+func InitVDEMode(config *conf.VDEConfig) *VDE {
+	mode := &VDE{
+		VDEConfig: config,
+	}
+
+	return mode
+}
+
+type VDE struct {
+	*conf.VDEConfig
+}
